A function that executes a function given as a parameter on each element of an array.

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -3,16 +3,16 @@
   * array_iterator - this is function
   * @array: param
   * @size: param2
+  * @action: param3
   */
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	long unsigned int i = 0;
+	size_t i = 0;
 	void (*f_ptr)(int) = action;
 
-	while(i != size)
+	while (i != size)
 	{
-		f_ptr (array[i]);
+		f_ptr(array[i]);
 		i++;
 	}
 }
-
